+++ src/test/java/org/stellar/sdk/operations/InvokeHostFunctionOperationTest.java
@@ -5,6 +5,8 @@ import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
+import java.math.BigInteger;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import org.junit.Test;
@@ -18,6 +20,7 @@ import org.stellar.sdk.xdr.ContractExecutableType;
 import org.stellar.sdk.xdr.ContractIDPreimage;
 import org.stellar.sdk.xdr.ContractIDPreimageType;
 import org.stellar.sdk.xdr.CreateContractArgs;
+import org.stellar.sdk.xdr.CreateContractArgsV2;
 import org.stellar.sdk.xdr.Hash;
 import org.stellar.sdk.xdr.HostFunction;
 import org.stellar.sdk.xdr.HostFunctionType;
@@ -322,11 +325,12 @@ public class InvokeHostFunctionOperationTest {
         };
     Address address = new Address("GAHJJJKMOKYE4RVPZEWZTKH5FVI4PA3VL7GK2LFNUBSGBV6OJP7TQSLX");
     InvokeHostFunctionOperation operation =
-        InvokeHostFunctionOperation.createContractOperationBuilder(wasmIdString, address, salt)
+        InvokeHostFunctionOperation.createContractOperationBuilder(
+                wasmIdString, address, null, salt)
             .build();
 
-    CreateContractArgs createContractArgs =
-        CreateContractArgs.builder()
+    CreateContractArgsV2 createContractArgs =
+        CreateContractArgsV2.builder()
             .contractIDPreimage(
                 ContractIDPreimage.builder()
                     .discriminant(ContractIDPreimageType.CONTRACT_ID_PREIMAGE_FROM_ADDRESS)
@@ -341,18 +345,19 @@ public class InvokeHostFunctionOperationTest {
                     .discriminant(ContractExecutableType.CONTRACT_EXECUTABLE_WASM)
                     .wasm_hash(new Hash(wasmId))
                     .build())
+            .constructorArgs(new SCVal[0])
             .build();
     HostFunction expectedFunction =
         HostFunction.builder()
-            .discriminant(HostFunctionType.HOST_FUNCTION_TYPE_CREATE_CONTRACT)
-            .createContract(createContractArgs)
+            .discriminant(HostFunctionType.HOST_FUNCTION_TYPE_CREATE_CONTRACT_V2)
+            .createContractV2(createContractArgs)
             .build();
 
     assertEquals(operation.getHostFunction(), expectedFunction);
     assertTrue(operation.getAuth().isEmpty());
     assertNull(operation.getSourceAccount());
     String expectedXdr =
-        "AAAAAAAAABgAAAABAAAAAAAAAAAAAAAADpSlTHKwTkavyS2ZqP0tUceDdV/MrSytoGRg185L/zgRMwIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHwAAAAAAAQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHwAAAAA=";
+        "AAAAAAAAABgAAAADAAAAAAAAAAAAAAAADpSlTHKwTkavyS2ZqP0tUceDdV/MrSytoGRg185L/zgRMwIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHwAAAAAAAQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHwAAAAAAAAAA";
     assertEquals(expectedXdr, operation.toXdrBase64());
   }
 
@@ -372,10 +377,11 @@ public class InvokeHostFunctionOperationTest {
         };
     Address address = new Address("GAHJJJKMOKYE4RVPZEWZTKH5FVI4PA3VL7GK2LFNUBSGBV6OJP7TQSLX");
     InvokeHostFunctionOperation operation =
-        InvokeHostFunctionOperation.createContractOperationBuilder(wasmId, address, salt).build();
+        InvokeHostFunctionOperation.createContractOperationBuilder(wasmId, address, null, salt)
+            .build();
 
-    CreateContractArgs createContractArgs =
-        CreateContractArgs.builder()
+    CreateContractArgsV2 createContractArgs =
+        CreateContractArgsV2.builder()
             .contractIDPreimage(
                 ContractIDPreimage.builder()
                     .discriminant(ContractIDPreimageType.CONTRACT_ID_PREIMAGE_FROM_ADDRESS)
@@ -390,18 +396,76 @@ public class InvokeHostFunctionOperationTest {
                     .discriminant(ContractExecutableType.CONTRACT_EXECUTABLE_WASM)
                     .wasm_hash(new Hash(wasmId))
                     .build())
+            .constructorArgs(new SCVal[0])
             .build();
     HostFunction expectedFunction =
         HostFunction.builder()
-            .discriminant(HostFunctionType.HOST_FUNCTION_TYPE_CREATE_CONTRACT)
-            .createContract(createContractArgs)
+            .discriminant(HostFunctionType.HOST_FUNCTION_TYPE_CREATE_CONTRACT_V2)
+            .createContractV2(createContractArgs)
             .build();
 
     assertEquals(operation.getHostFunction(), expectedFunction);
     assertTrue(operation.getAuth().isEmpty());
     assertNull(operation.getSourceAccount());
     String expectedXdr =
-        "AAAAAAAAABgAAAABAAAAAAAAAAAAAAAADpSlTHKwTkavyS2ZqP0tUceDdV/MrSytoGRg185L/zgRMwIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHwAAAAAAAQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHwAAAAA=";
+        "AAAAAAAAABgAAAADAAAAAAAAAAAAAAAADpSlTHKwTkavyS2ZqP0tUceDdV/MrSytoGRg185L/zgRMwIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHwAAAAAAAQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHwAAAAAAAAAA";
+    assertEquals(expectedXdr, operation.toXdrBase64());
+  }
+
+  @Test
+  public void createContractOperationBuilderWithConstructorArgs() {
+    byte[] wasmId =
+        new byte[] {
+          0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e,
+          0x0f, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d,
+          0x1e, 0x1f
+        };
+    String wasmIdString = Util.bytesToHex(wasmId);
+    byte[] salt =
+        new byte[] {
+          0x11, 0x33, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e,
+          0x0f, 0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d,
+          0x1e, 0x1f
+        };
+    Address address = new Address("GAHJJJKMOKYE4RVPZEWZTKH5FVI4PA3VL7GK2LFNUBSGBV6OJP7TQSLX");
+    List<SCVal> constructorArgs =
+        Arrays.asList(
+            Scv.toAddress("GA2KQTETIRREL66P64GV6KCVICPULLDVHWJDZSIJKDLIAGBXUCIZ6P6E"),
+            Scv.toUint64(BigInteger.valueOf(123456789L)));
+    InvokeHostFunctionOperation operation =
+        InvokeHostFunctionOperation.createContractOperationBuilder(
+                wasmIdString, address, constructorArgs, salt)
+            .build();
+
+    CreateContractArgsV2 createContractArgs =
+        CreateContractArgsV2.builder()
+            .contractIDPreimage(
+                ContractIDPreimage.builder()
+                    .discriminant(ContractIDPreimageType.CONTRACT_ID_PREIMAGE_FROM_ADDRESS)
+                    .fromAddress(
+                        ContractIDPreimage.ContractIDPreimageFromAddress.builder()
+                            .address(address.toSCAddress())
+                            .salt(new Uint256(salt))
+                            .build())
+                    .build())
+            .executable(
+                ContractExecutable.builder()
+                    .discriminant(ContractExecutableType.CONTRACT_EXECUTABLE_WASM)
+                    .wasm_hash(new Hash(wasmId))
+                    .build())
+            .constructorArgs(constructorArgs.toArray(new SCVal[0]))
+            .build();
+    HostFunction expectedFunction =
+        HostFunction.builder()
+            .discriminant(HostFunctionType.HOST_FUNCTION_TYPE_CREATE_CONTRACT_V2)
+            .createContractV2(createContractArgs)
+            .build();
+
+    assertEquals(operation.getHostFunction(), expectedFunction);
+    assertTrue(operation.getAuth().isEmpty());
+    assertNull(operation.getSourceAccount());
+    String expectedXdr =
+        "AAAAAAAAABgAAAADAAAAAAAAAAAAAAAADpSlTHKwTkavyS2ZqP0tUceDdV/MrSytoGRg185L/zgRMwIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHwAAAAAAAQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRobHB0eHwAAAAIAAAASAAAAAAAAAAA0qEyTRGJF+8/3DV8oVUCfRax1PZI8yQlQ1oAYN6CRnwAAAAUAAAAAB1vNFQAAAAA=";
     assertEquals(expectedXdr, operation.toXdrBase64());
   }